Write code for the console version of this game
Wrote entire code for the console version of this game.
First, I created a function that would randomly select computer's
choice. Moving on, I then defined a function that returned
the result of one round.

Finally, I created one more function that took the above function
and iterated over it 5 times to play 5 rounds and then kept track
of each of their scores.